+++ example/lib/main.dart
@@ -1,6 +1,13 @@
-import 'package:flutter/material.dart';
+import 'dart:io';
 
-void main() => runApp(MyApp());
+import 'package:flutter/material.dart';
+import 'package:flutter_store_listing/flutter_store_listing.dart';
+import 'package:logging_appenders/logging_appenders.dart';
+
+void main() {
+  PrintAppender.setupLogging();
+  runApp(MyApp());
+}
 
 class MyApp extends StatefulWidget {
   @override
@@ -8,9 +15,19 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+  String _appId = 'unknown';
+  final _fsl = FlutterStoreListing();
+
   @override
   void initState() {
     super.initState();
+    (() async {
+      if (Platform.isIOS) {
+        _appId = await _fsl.getIosAppId() ?? 'app id not found';
+      } else if (Platform.isAndroid) {
+        _appId = await _fsl.getAndroidPackageName();
+      }
+    })();
   }
 
   @override
@@ -20,8 +37,16 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: const Center(
-          child: Text('Running on'),
+        body: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: <Widget>[
+            Text(_appId),
+            RaisedButton(
+                child: const Text('Open Store Listing'),
+                onPressed: () async {
+                  await _fsl.launchStoreListing();
+                }),
+          ],
         ),
       ),
     );