+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.1
+
+* More documentation, updated links.
+* No longer a singleton, override app store id and android package name.
+* Add isSupported() methods.
+* Made API more consistent.
+
 ## 0.1.0
 
 * First release